Allow to have lectures in the interface, ie non-interactive exercises Use it for the lesson chooser instead of a separate frame
This patch is a bit intrusive, sorry about it, but adding a superclass
to one of the most overloaded class of our hierarchie wasn't easy. I
should have split it in parts, in a better world, but life is crap
sometimes.

The good point is that this feature will help solving a lot of issues,
such as slow loading (load only one lesson at a time), or difficulties
for teacher to cume up with their own lessons.

+++ src/jlm/core/model/FileSessionKit.java
@@ -8,6 +8,7 @@ import java.io.FileWriter;
 import java.io.IOException;
 
 import jlm.core.model.lesson.Exercise;
+import jlm.core.model.lesson.Lecture;
 import jlm.core.model.lesson.Lesson;
 import jlm.core.model.lesson.SourceFile;
 import jlm.core.model.lesson.SourceFileAliased;
@@ -38,74 +39,76 @@ public class FileSessionKit /* FIXME implements ISessionKit  */ {
 				Logger.log("FileSessionKit:store", "cannot create session store directory");
 			};
 
-		// File lessonDir;
-		File exerciseDir;
-		for (Lesson lesson : this.game.getLessons()) {
-			// lessonDir = new File(saveDir + sep +
-			// lesson.getClass().getName());
-			// if (!lessonDir.exists())
-			// lessonDir.mkdir();
-			// // save lesson data
+			// File lessonDir;
+			File exerciseDir;
+			for (Lesson lesson : this.game.getLessons()) {
+				// lessonDir = new File(saveDir + sep +
+				// lesson.getClass().getName());
+				// if (!lessonDir.exists())
+				// lessonDir.mkdir();
+				// // save lesson data
 
-			for (Exercise exercise : lesson.exercises()) {
-				exerciseDir = new File(SAVE_DIR, exercise.getClass().getName());
-				if (!exerciseDir.exists())
-					if (! exerciseDir.mkdir()) 
-						Logger.log("FileSessionKit:store", "cannot remove "+exerciseDir+" directory");
+				for (Lecture lecture : lesson.exercises()) {
+					if (lecture instanceof Exercise) {
+						Exercise exercise = (Exercise) lecture;
+						exerciseDir = new File(SAVE_DIR, exercise.getClass().getName());
+						if (!exerciseDir.exists())
+							if (! exerciseDir.mkdir()) 
+								Logger.log("FileSessionKit:store", "cannot remove "+exerciseDir+" directory");
 
 
-				// create file DONE if exercise has been successfully passed
-				File exerciseFile = new File(exerciseDir, "DONE");
-				if (exercise.isSuccessfullyPassed()) {
-					if (!exerciseFile.exists()) {
-						try {
-							exerciseFile.createNewFile();
-						} catch (IOException e) {
-							e.printStackTrace();
-						}
-					}
-				} else {
-					if (exerciseFile.exists()) {
-						if (!exerciseFile.delete()) {
-							Logger.log("FileSessionKit:store", "cannot remove "+exerciseFile+" directory");
-						}
-					}
-				}
-
-				// save exercise body
-				for (ProgrammingLanguage lang:exercise.getProgLanguages()) 
-					for (int i = 0; i < exercise.publicSourceFileCount(lang); i++) {
-						SourceFile sf = exercise.getPublicSourceFile(lang,i);
-
-						if (!(sf instanceof SourceFileRevertable))
-							continue;
-
-						SourceFileRevertable srcFile = (SourceFileRevertable)sf;
-						File outputFile = new File(exerciseDir+"/"+lang, srcFile.getName());
-
-						if (srcFile.hasChanged()) {
-							BufferedWriter bw = null;
-							try {
-								bw = new BufferedWriter(new FileWriter(outputFile));
-								bw.write(srcFile.getBody());
-							} catch (IOException e) {
-								e.printStackTrace();
-							} finally {
+						// create file DONE if exercise has been successfully passed
+						File exerciseFile = new File(exerciseDir, "DONE");
+						if (exercise.isSuccessfullyPassed()) {
+							if (!exerciseFile.exists()) {
 								try {
-									bw.close();
+									exerciseFile.createNewFile();
 								} catch (IOException e) {
 									e.printStackTrace();
 								}
 							}
 						} else {
-							if (outputFile.exists())
-								if (! outputFile.delete())
-									Logger.log("FileSessionKit:store", "cannot remove "+outputFile);
+							if (exerciseFile.exists()) {
+								if (!exerciseFile.delete()) {
+									Logger.log("FileSessionKit:store", "cannot remove "+exerciseFile+" directory");
+								}
+							}
 						}
-					}
-				
-			} // end-for exercise
-		} // end-for lesson
+
+						// save exercise body
+						for (ProgrammingLanguage lang:exercise.getProgLanguages()) 
+							for (int i = 0; i < exercise.publicSourceFileCount(lang); i++) {
+								SourceFile sf = exercise.getPublicSourceFile(lang,i);
+
+								if (!(sf instanceof SourceFileRevertable))
+									continue;
+
+								SourceFileRevertable srcFile = (SourceFileRevertable)sf;
+								File outputFile = new File(exerciseDir+"/"+lang, srcFile.getName());
+
+								if (srcFile.hasChanged()) {
+									BufferedWriter bw = null;
+									try {
+										bw = new BufferedWriter(new FileWriter(outputFile));
+										bw.write(srcFile.getBody());
+									} catch (IOException e) {
+										e.printStackTrace();
+									} finally {
+										try {
+											bw.close();
+										} catch (IOException e) {
+											e.printStackTrace();
+										}
+									}
+								} else {
+									if (outputFile.exists())
+										if (! outputFile.delete())
+											Logger.log("FileSessionKit:store", "cannot remove "+outputFile);
+								}
+							}
+					} // is exercise
+				} // end-for lecture
+			} // end-for lesson
 	}
 
 	public void loadAll() {
@@ -115,53 +118,56 @@ public class FileSessionKit /* FIXME implements ISessionKit  */ {
 		// File lessonDir;
 		File exerciseDir;
 		for (Lesson lesson : this.game.getLessons()) {
-			for (Exercise exercise : lesson.exercises()) {
-				exerciseDir = new File(SAVE_DIR, exercise.getClass().getName());
-				if (!exerciseDir.exists())
-					continue;
+			for (Lecture lecture : lesson.exercises()) {
+				if (lecture instanceof Exercise) {
+					Exercise exercise = (Exercise) lecture;
+					exerciseDir = new File(SAVE_DIR, exercise.getClass().getName());
+					if (!exerciseDir.exists())
+						continue;
 
-				File exerciseFile = new File(exerciseDir, "DONE");
-				if (exerciseFile.exists()) {
-					exercise.successfullyPassed();
-				}
+					File exerciseFile = new File(exerciseDir, "DONE");
+					if (exerciseFile.exists()) {
+						exercise.successfullyPassed();
+					}
 
-				// load exercise body
-				for (ProgrammingLanguage lang:exercise.getProgLanguages()) {
-					for (int i = 0; i < exercise.publicSourceFileCount(lang); i++) {
-						SourceFile srcFile = exercise.getPublicSourceFile(lang,i);
+					// load exercise body
+					for (ProgrammingLanguage lang:exercise.getProgLanguages()) {
+						for (int i = 0; i < exercise.publicSourceFileCount(lang); i++) {
+							SourceFile srcFile = exercise.getPublicSourceFile(lang,i);
 
-						if (srcFile instanceof SourceFileAliased)
-							continue;
+							if (srcFile instanceof SourceFileAliased)
+								continue;
 
-						File of = new File(exerciseDir+"/"+lang, srcFile.getName());
-						if (!of.exists()) /* try to load using the old format (not specifying the programming language) */
-							of = new File(exerciseDir+"/"+lang, srcFile.getName());
-						if (of.exists()) {
-							BufferedReader br = null;
-							try {
-								br = new BufferedReader(new FileReader(of));
-								String s;
-								StringBuffer b = new StringBuffer();
-
-								while ((s = br.readLine()) != null) {
-									b.append(s);
-									b.append("\n");
-								}
-
-								srcFile.setBody(b.toString());
-							} catch (IOException e) {
-								e.printStackTrace();
-							} finally {
+							File of = new File(exerciseDir+"/"+lang, srcFile.getName());
+							if (!of.exists()) /* try to load using the old format (not specifying the programming language) */
+								of = new File(exerciseDir+"/"+lang, srcFile.getName());
+							if (of.exists()) {
+								BufferedReader br = null;
 								try {
-									br.close();
+									br = new BufferedReader(new FileReader(of));
+									String s;
+									StringBuffer b = new StringBuffer();
+
+									while ((s = br.readLine()) != null) {
+										b.append(s);
+										b.append("\n");
+									}
+
+									srcFile.setBody(b.toString());
 								} catch (IOException e) {
 									e.printStackTrace();
+								} finally {
+									try {
+										br.close();
+									} catch (IOException e) {
+										e.printStackTrace();
+									}
 								}
 							}
 						}
 					}
-				}
-			} // end-for exercise
+				} // is exercise
+			} // end-for lecture
 		} // end-for lesson
 	}
 
@@ -171,34 +177,37 @@ public class FileSessionKit /* FIXME implements ISessionKit  */ {
 
 		File exerciseDir;
 		for (Lesson lesson : this.game.getLessons()) {
-			for (Exercise exercise : lesson.exercises()) {
-				exerciseDir = new File(SAVE_DIR, exercise.getClass().getName());
-				if (!exerciseDir.exists())
-					continue;
+			for (Lecture lecture : lesson.exercises()) {
+				if (lecture instanceof Exercise) {
+					Exercise exercise = (Exercise) lecture;
+					exerciseDir = new File(SAVE_DIR, exercise.getClass().getName());
+					if (!exerciseDir.exists())
+						continue;
 
-				File exerciseFile = new File(exerciseDir, "DONE");
-				if (exerciseFile.exists()) {
-					if (!exerciseFile.delete()) {
-						Logger.log("FileSessionKit:cleanUp", "cannot remove "+exerciseFile);
-					};
-				}
-
-				for (ProgrammingLanguage lang:exercise.getProgLanguages()) 
-					for (int i = 0; i < exercise.publicSourceFileCount(lang); i++) {
-						SourceFile srcFile = exercise.getPublicSourceFile(lang, i);
-
-						File of = new File(exerciseDir, srcFile.getName());
-						if (of.exists()) {
-							if (! of.delete()) {
-								Logger.log("FileSessionKit:cleanUp", "cannot remove "+exerciseDir+"/"+srcFile.getName()+" directory");				
-							}
-						}
+					File exerciseFile = new File(exerciseDir, "DONE");
+					if (exerciseFile.exists()) {
+						if (!exerciseFile.delete()) {
+							Logger.log("FileSessionKit:cleanUp", "cannot remove "+exerciseFile);
+						};
 					}
 
-				if (! exerciseDir.delete()) {
-					Logger.log("FileSessionKit:cleanUp", "cannot remove "+exerciseDir+" directory");					
-				}
-			} // end-for exercise
+					for (ProgrammingLanguage lang:exercise.getProgLanguages()) 
+						for (int i = 0; i < exercise.publicSourceFileCount(lang); i++) {
+							SourceFile srcFile = exercise.getPublicSourceFile(lang, i);
+
+							File of = new File(exerciseDir, srcFile.getName());
+							if (of.exists()) {
+								if (! of.delete()) {
+									Logger.log("FileSessionKit:cleanUp", "cannot remove "+exerciseDir+"/"+srcFile.getName()+" directory");				
+								}
+							}
+						}
+
+					if (! exerciseDir.delete()) {
+						Logger.log("FileSessionKit:cleanUp", "cannot remove "+exerciseDir+" directory");					
+					}
+				} // is exercise
+			} // end-for lectures
 		} // end-for lesson
 		if (! SAVE_DIR.delete()) {
 			Logger.log("FileSessionKit:cleanup", "cannot remove session store directory");
